// types/starling/feedItems.ts
type CurrencyAndAmount = {
  currency: Currency;
  minorUnits: number;
};

export enum Direction {
  IN = "IN",
  OUT = "OUT"
}

export enum Source {
  BRITISH_BUSINESS_BANK_FEES = "BRITISH_BUSINESS_BANK_FEES",
  CARD_FEE_CHARGE = "CARD_FEE_CHARGE",
  CASH_DEPOSIT = "CASH_DEPOSIT",
  CASH_DEPOSIT_CHARGE = "CASH_DEPOSIT_CHARGE",
  CASH_WITHDRAWAL = "CASH_WITHDRAWAL",
  CASH_WITHDRAWAL_CHARGE = "CASH_WITHDRAWAL_CHARGE",
  CHAPS = "CHAPS",
  CHEQUE = "CHEQUE",
  CICS_CHEQUE = "CICS_CHEQUE",
  CURRENCY_CLOUD = "CURRENCY_CLOUD",
  DIRECT_CREDIT = "DIRECT_CREDIT",
  DIRECT_DEBIT = "DIRECT_DEBIT",
  DIRECT_DEBIT_DISPUTE = "DIRECT_DEBIT_DISPUTE",
  INTERNAL_TRANSFER = "INTERNAL_TRANSFER",
  MASTER_CARD = "MASTER_CARD",
  MASTERCARD_MONEYSEND = "MASTERCARD_MONEYSEND",
  MASTERCARD_CHARGEBACK = "MASTERCARD_CHARGEBACK",
  MISSED_PAYMENT_FEE = "MISSED_PAYMENT_FEE",
  FASTER_PAYMENTS_IN = "FASTER_PAYMENTS_IN",
  FASTER_PAYMENTS_OUT = "FASTER_PAYMENTS_OUT",
  FASTER_PAYMENTS_REVERSAL = "FASTER_PAYMENTS_REVERSAL",
  STRIPE_FUNDING = "STRIPE_FUNDING",
  INTEREST_PAYMENT = "INTEREST_PAYMENT",
  NOSTRO_DEPOSIT = "NOSTRO_DEPOSIT",
  OVERDRAFT = "OVERDRAFT",
  OVERDRAFT_INTEREST_WAIVED = "OVERDRAFT_INTEREST_WAIVED",
  FASTER_PAYMENTS_REFUND = "FASTER_PAYMENTS_REFUND",
  STARLING_PAY_STRIPE = "STARLING_PAY_STRIPE",
  ON_US_PAY_ME = "ON_US_PAY_ME",
  LOAN_PRINCIPAL_PAYMENT = "LOAN_PRINCIPAL_PAYMENT",
  LOAN_REPAYMENT = "LOAN_REPAYMENT",
  LOAN_OVERPAYMENT = "LOAN_OVERPAYMENT",
  LOAN_LATE_PAYMENT = "LOAN_LATE_PAYMENT",
  LOAN_FEE_PAYMENT = "LOAN_FEE_PAYMENT",
  LOAN_INTEREST_CHARGE = "LOAN_INTEREST_CHARGE",
  SEPA_CREDIT_TRANSFER = "SEPA_CREDIT_TRANSFER",
  SEPA_DIRECT_DEBIT = "SEPA_DIRECT_DEBIT",
  TARGET2_CUSTOMER_PAYMENT = "TARGET2_CUSTOMER_PAYMENT",
  FX_TRANSFER = "FX_TRANSFER",
  ISS_PAYMENT = "ISS_PAYMENT",
  STARLING_PAYMENT = "STARLING_PAYMENT",
  SUBSCRIPTION_CHARGE = "SUBSCRIPTION_CHARGE",
  OVERDRAFT_FEE = "OVERDRAFT_FEE",
  WITHHELD_TAX = "WITHHELD_TAX",
  ERRORS_AND_OMISSIONS = "ERRORS_AND_OMISSIONS",
  OWN_FUNDS_TOPUP = "OWN_FUNDS_TOPUP",
  INTEREST_V2_PAYMENT = "INTEREST_V2_PAYMENT"
}

export enum SubSource {
  CONTACTLESS = "CONTACTLESS",
  MAGNETIC_STRIP = "MAGNETIC_STRIP",
  MANUAL_KEY_ENTRY = "MANUAL_KEY_ENTRY",
  CHIP_AND_PIN = "CHIP_AND_PIN",
  ONLINE = "ONLINE",
  ATM = "ATM",
  CREDIT_AUTH = "CREDIT_AUTH",
  APPLE_PAY = "APPLE_PAY",
  APPLE_PAY_ONLINE = "APPLE_PAY_ONLINE",
  ANDROID_PAY = "ANDROID_PAY",
  ANDROID_PAY_ONLINE = "ANDROID_PAY_ONLINE",
  FITBIT_PAY = "FITBIT_PAY",
  GARMIN_PAY = "GARMIN_PAY",
  SAMSUNG_PAY = "SAMSUNG_PAY",
  OTHER_WALLET = "OTHER_WALLET",
  CARD_SUBSCRIPTION = "CARD_SUBSCRIPTION",
  NOT_APPLICABLE = "NOT_APPLICABLE",
  UNKNOWN = "UNKNOWN",
  DEPOSIT = "DEPOSIT",
  OVERDRAFT = "OVERDRAFT",
  SETTLE_UP = "SETTLE_UP",
  NEARBY = "NEARBY",
  TRANSFER_SAME_CURRENCY = "TRANSFER_SAME_CURRENCY",
  NEW_CARD = "NEW_CARD"
}

export enum Status {
  UPCOMING = "UPCOMING",
  PENDING = "PENDING",
  REVERSED = "REVERSED",
  SETTLED = "SETTLED",
  DECLINED = "DECLINED",
  REFUNDED = "REFUNDED",
  RETRYING = "RETRYING",
  ACCOUNT_CHECK = "ACCOUNT_CHECK"
}

export enum CounterParty {
  CATEGORY = "CATEGORY",
  CHEQUE = "CHEQUE",
  CUSTOMER = "CUSTOMER",
  PAYEE = "PAYEE",
  MERCHANT = "MERCHANT",
  SENDER = "SENDER",
  STARLING = "STARLING",
  LOAN = "LOAN"
}

export enum Country {
  AC = "AC",
  AD = "AD",
  AE = "AE",
  AF = "AF",
  AG = "AG",
  AI = "AI",
  AL = "AL",
  AM = "AM",
  AN = "AN",
  AO = "AO",
  AQ = "AQ",
  AR = "AR",
  AS = "AS",
  AT = "AT",
  AU = "AU",
  AW = "AW",
  AX = "AX",
  AZ = "AZ",
  BA = "BA",
  BB = "BB",
  BD = "BD",
  BE = "BE",
  BF = "BF",
  BG = "BG",
  BH = "BH",
  BI = "BI",
  BJ = "BJ",
  BL = "BL",
  BM = "BM",
  BN = "BN",
  BO = "BO",
  BQ = "BQ",
  BR = "BR",
  BS = "BS",
  BT = "BT",
  BU = "BU",
  BV = "BV",
  BW = "BW",
  BY = "BY",
  BZ = "BZ",
  CA = "CA",
  CC = "CC",
  CD = "CD",
  CF = "CF",
  CG = "CG",
  CH = "CH",
  CI = "CI",
  CK = "CK",
  CL = "CL",
  CM = "CM",
  CN = "CN",
  CO = "CO",
  CP = "CP",
  CR = "CR",
  CS = "CS",
  CU = "CU",
  CV = "CV",
  CW = "CW",
  CX = "CX",
  CY = "CY",
  CZ = "CZ",
  DE = "DE",
  DG = "Dg",
  DJ = "DJ",
  DK = "DK",
  DM = "DM",
  DO = "DO",
  DZ = "DZ",
  EA = "EA",
  EC = "EC",
  EE = "EE",
  EG = "EG",
  EH = "EH",
  ER = "ER",
  ES = "ES",
  ET = "ET",
  EU = "EU",
  EZ = "EZ",
  FI = "FI",
  FJ = "FJ",
  FK = "FK",
  FM = "FM",
  FO = "FO",
  FR = "FR",
  FX = "FX",
  GA = "GA",
  GB = "GB",
  GD = "GD",
  GE = "GE",
  GF = "GF",
  GG = "GG",
  GH = "GH",
  GI = "GI",
  GL = "GL",
  GM = "GM",
  GN = "GN",
  GP = "GP",
  GQ = "GQ",
  GR = "GR",
  GS = "GS",
  GT = "GT",
  GU = "GU",
  GW = "GW",
  GY = "GY",
  HK = "HK",
  HM = "HM",
  HN = "HN",
  HR = "HR",
  HT = "HT",
  HU = "HU",
  IC = "IC",
  ID = "ID",
  IE = "IE",
  IL = "IL",
  IM = "IM",
  IN = "IN",
  IO = "IO",
  IQ = "IQ",
  IR = "IR",
  IS = "IS",
  IT = "IT",
  JE = "JE",
  JM = "JM",
  JO = "JO",
  JP = "JP",
  KE = "KE",
  KG = "KG",
  KH = "KH",
  KI = "KI",
  KM = "KM",
  KN = "KN",
  KP = "KP",
  KR = "KR",
  KW = "KW",
  KY = "KY",
  KZ = "KZ",
  LA = "LA",
  LB = "LB",
  LC = "LC",
  LI = "LI",
  LK = "LK",
  LR = "LR",
  LS = "LS",
  LT = "LT",
  LU = "LU",
  LV = "LV",
  LY = "LY",
  MA = "MA",
  MC = "MC",
  MD = "MD",
  ME = "ME",
  MF = "MF",
  MG = "MG",
  MH = "MH",
  MK = "MK",
  ML = "ML",
  MM = "MM",
  MN = "MN",
  MO = "MO",
  MP = "MP",
  MQ = "MQ",
  MR = "MR",
  MS = "MS",
  MT = "MT",
  MU = "MU",
  MV = "MV",
  MW = "MW",
  MX = "MX",
  MY = "MY",
  MZ = "MZ",
  NA = "NA",
  NC = "NC",
  NE = "NE",
  NF = "NF",
  NG = "NG",
  NI = "NI",
  NL = "NL",
  NO = "NO",
  NP = "NP",
  NR = "NR",
  NT = "NT",
  NU = "NU",
  NZ = "NZ",
  OM = "OM",
  PA = "PA",
  PE = "PE",
  PF = "PF",
  PG = "PG",
  PH = "PH",
  PK = "PK",
  PL = "PL",
  PM = "PM",
  PN = "PN",
  PR = "PR",
  PS = "PS",
  PT = "PT",
  PW = "PW",
  PY = "PY",
  QA = "QA",
  RE = "RE",
  RO = "RO",
  RS = "RS",
  RU = "RU",
  RW = "RW",
  SA = "SA",
  SB = "SB",
  SC = "SC",
  SD = "SD",
  SE = "SE",
  SF = "SF",
  SG = "SG",
  SH = "SH",
  SI = "SI",
  SJ = "SJ",
  SK = "SK",
  SL = "SL",
  SM = "SM",
  SN = "SN",
  SO = "SO",
  SR = "SR",
  SS = "SS",
  ST = "ST",
  SU = "SU",
  SV = "SV",
  SX = "SX",
  SY = "SY",
  SZ = "SZ",
  TA = "TA",
  TC = "TC",
  TD = "TD",
  TF = "TF",
  TG = "TG",
  TH = "TH",
  TJ = "TJ",
  TK = "TK",
  TL = "TL",
  TM = "TM",
  TN = "TN",
  TO = "TO",
  TP = "TP",
  TR = "TR",
  TT = "TT",
  TV = "TV",
  TW = "TW",
  TZ = "TZ",
  UA = "UA",
  UG = "UG",
  UK = "UK",
  UM = "UM",
  US = "US",
  UY = "UY",
  UZ = "UZ",
  VA = "VA",
  VC = "VC",
  VE = "VE",
  VG = "VG",
  VI = "VI",
  VN = "VN",
  VU = "VU",
  WF = "WF",
  WS = "WS",
  XI = "XI",
  XU = "XU",
  XK = "XK",
  XX = "XX",
  YE = "YE",
  YT = "YT",
  YU = "YU",
  ZA = "ZA",
  ZM = "ZM",
  ZR = "ZR",
  ZW = "ZW",
  UNDEFINED = "UNDEFINED"
}

export enum SpendingCategory {
  BIKE = "BIKE",
  BILLS_AND_SERVICES = "BILLS_AND_SERVICES",
  BUCKET_LIST = "BUCKET_LIST",
  CAR = "CAR",
  CASH = "CASH",
  CELEBRATION = "CELEBRATION",
  CHARITY = "CHARITY",
  CHILDREN = "CHILDREN",
  CLOTHES = "CLOTHES",
  COFFEE = "COFFEE",
  DEBT_REPAYMENT = "DEBT_REPAYMENT",
  DIY = "DIY",
  DRINKS = "DRINKS",
  EATING_OUT = "EATING_OUT",
  EDUCATION = "EDUCATION",
  EMERGENCY = "EMERGENCY",
  ENTERTAINMENT = "ENTERTAINMENT",
  ESSENTIAL_SPEND = "ESSENTIAL_SPEND",
  EXPENSES = "EXPENSES",
  FAMILY = "FAMILY",
  FITNESS = "FITNESS",
  FUEL = "FUEL",
  GAMBLING = "GAMBLING",
  GAMING = "GAMING",
  GARDEN = "GARDEN",
  GENERAL = "GENERAL",
  GIFTS = "GIFTS",
  GROCERIES = "GROCERIES",
  HOBBY = "HOBBY",
  HOLIDAYS = "HOLIDAYS",
  HOME = "HOME",
  IMPULSE_BUY = "IMPULSE_BUY",
  INCOME = "INCOME",
  INSURANCE = "INSURANCE",
  INVESTMENTS = "INVESTMENTS",
  LIFESTYLE = "LIFESTYLE",
  MAINTENANCE_AND_REPAIRS = "MAINTENANCE_AND_REPAIRS",
  MEDICAL = "MEDICAL",
  MORTGAGE = "MORTGAGE",
  NON_ESSENTIAL_SPEND = "NON_ESSENTIAL_SPEND",
  PAYMENTS = "PAYMENTS",
  PERSONAL_CARE = "PERSONAL_CARE",
  PERSONAL_TRANSFERS = "PERSONAL_TRANSFERS",
  PETS = "PETS",
  PROJECTS = "PROJECTS",
  RELATIONSHIPS = "RELATIONSHIPS",
  RENT = "RENT",
  SAVING = "SAVING",
  SHOPPING = "SHOPPING",
  SUBSCRIPTIONS = "SUBSCRIPTIONS",
  TAKEAWAY = "TAKEAWAY",
  TAXI = "TAXI",
  TRANSPORT = "TRANSPORT",
  TREATS = "TREATS",
  WEDDING = "WEDDING",
  WELLBEING = "WELLBEING",
  NONE = "NONE",
  REVENUE = "REVENUE",
  OTHER_INCOME = "OTHER_INCOME",
  CLIENT_REFUNDS = "CLIENT_REFUNDS",
  INVENTORY = "INVENTORY",
  STAFF = "STAFF",
  TRAVEL = "TRAVEL",
  WORKPLACE = "WORKPLACE",
  REPAIRS_AND_MAINTENANCE = "REPAIRS_AND_MAINTENANCE",
  ADMIN = "ADMIN",
  MARKETING = "MARKETING",
  BUSINESS_ENTERTAINMENT = "BUSINESS_ENTERTAINMENT",
  INTEREST_PAYMENTS = "INTEREST_PAYMENTS",
  BANK_CHARGES = "BANK_CHARGES",
  OTHER = "OTHER",
  FOOD_AND_DRINK = "FOOD_AND_DRINK",
  EQUIPMENT = "EQUIPMENT",
  PROFESSIONAL_SERVICES = "PROFESSIONAL_SERVICES",
  PHONE_AND_INTERNET = "PHONE_AND_INTERNET",
  VEHICLES = "VEHICLES",
  DIRECTORS_WAGES = "DIRECTORS_WAGES",
  VAT = "VAT",
  CORPORATION_TAX = "CORPORATION_TAX",
  SELF_ASSESSMENT_TAX = "SELF_ASSESSMENT_TAX",
  INVESTMENT_CAPITAL = "INVESTMENT_CAPITAL",
  TRANSFERS = "TRANSFERS",
  LOAN_PRINCIPAL = "LOAN_PRINCIPAL",
  PERSONAL = "PERSONAL",
  DIVIDENDS = "DIVIDENDS"
}

type AssociatedFeeRoundUp = {
  goalCategoryUid: string;
  amount: CurrencyAndAmount;
};

export enum BatchPaymentType {
  BULK_PAYMENT = "BULK_PAYMENT"
}

type BatchPaymentDetails = {
  batchPaymentUid: string;
  batchPaymentType: BatchPaymentType;
};

export type StarlingFeedItem = {
  feedItemUid: string;
  categoryUid: string;
  amount: CurrencyAndAmount;
  sourceAmount: CurrencyAndAmount;
  direction: Direction;
  updatedAt: string;
  transactionTime: string;
  settlementTime: string;
  retryAllocationUntilTime: string;
  source: Source;
  sourceSubType: SubSource;
  status: Status;
  transactingApplicationUserUid: string;
  counterPartyType: CounterParty;
  counterPartyUid: string;
  counterPartyName: string;
  counterPartySubEntityUid: string;
  counterPartySubEntityName: string;
  counterPartySubEntityIdentifier: string;
  exchangeRate: number;
  totalFees: number;
  totalFeeAmount: CurrencyAndAmount;
  reference: string;
  country: Country;
  spendingCategory: SpendingCategory;
  userNote: string;
  roundUp: AssociatedFeeRoundUp;
  hasAttachment: boolean;
  hasReceipt: boolean;
  batchPaymentDetails: BatchPaymentDetails;
};